moved canvas ref to universe
